requirements.txt
pyspark
confluent-kafka[avro]
fastavro
python-dotenv
requests
